test(agent): compare root exports by value and split the declaration pin
Review of #2257, two findings.

The root check compared NAMES, so an alias passed it: re-export
`systemRecordRequesterLaneEnabledV1 as resolveSystemRecordRequesterLaneV1` from
the barrel and the internal function is published under a name the loop never
looks for. The contract is that these functions stay off the root, not that
these spellings do, so the comparison is now by VALUE and the error names both
the internal export and the alias it was published under.

The root TYPE pin also sat in a source-level coverage file, which hid a real
dependency: that suite otherwise reads `../src/...`, while the root import
resolves through generated `dist/index.d.ts`, so `test:types` alone on a
never-built checkout would not resolve it. Split into its own
root-surface typecheck whose header states the ordering rather than implying it
-- `build` is `tsc && test:types && test:package-root`, so the emit always
precedes it. The assertion cannot move to the source module: a source import
keeps passing while the barrel drops the type, which is the drift it pins.

Refs #2052.
