+++ packages/agent/test/system-record-config-controls-coverage.typecheck.ts
@@ -47,52 +47,3 @@ type NotDeclared = Exclude<keyof SystemRecordConfigControlsV1, ControlKeysOnAgen
 
 export type _EveryAgentControlIsForwarded = Expect<IsEmpty<NotForwarded>>;
 export type _EveryForwardedControlIsDeclared = Expect<IsEmpty<NotDeclared>>;
-
-/**
- * Root-level TYPE contract, which the runtime package-root script cannot reach:
- * TypeScript types are erased, so `test-package-root.mjs` can prove the picker
- * is exported from the package root and prove the four resolvers are not, but
- * it is blind to whether `SystemRecordConfigControlsV1` still comes with it.
- * Dropping the type export while keeping the picker would leave that script
- * green and break every consumer that names the picker's return type.
- *
- * These imports deliberately come from the PACKAGE ROOT rather than the source
- * module the rest of this file uses: the assertions above are about the code,
- * this one is about the published surface, and only a root import can fail when
- * the barrel changes.
- */
-import type {
-  pickSystemRecordConfigControlsV1 as rootPickSystemRecordConfigControlsV1,
-  SystemRecordConfigControlsV1 as RootSystemRecordConfigControlsV1,
-} from '@origintrail-official/dkg-agent';
-
-type RootPickerReturn = ReturnType<typeof rootPickSystemRecordConfigControlsV1>;
-
-/**
- * KEY-SET equality, not mutual assignability. Every control field is optional,
- * and two object types whose keys differ only by optional members are mutually
- * assignable — so a pair of assignment checks would pass while the root's
- * published type gained a control the picker never forwards, or lost one it
- * does. That is a pin that cannot fail for the drift it exists to catch, which
- * is why this reuses the `Exclude<keyof A, keyof B>` shape above rather than
- * asserting the two types are compatible.
- */
-type PublishedButNotReturned = Exclude<
-  keyof RootSystemRecordConfigControlsV1,
-  keyof RootPickerReturn
->;
-type ReturnedButNotPublished = Exclude<
-  keyof RootPickerReturn,
-  keyof RootSystemRecordConfigControlsV1
->;
-
-export type _RootTypePublishesExactlyWhatThePickerReturns = Expect<
-  IsEmpty<PublishedButNotReturned>
->;
-export type _RootPickerReturnsExactlyWhatTheRootTypePublishes = Expect<
-  IsEmpty<ReturnedButNotPublished>
->;
-
-/** The published type must still describe the picker's value, not merely share keys. */
-export type _RootTypeIsAssignableFromThePickerReturn =
-  Expect<RootPickerReturn extends RootSystemRecordConfigControlsV1 ? true : false>;
